// backend/src/domain/notifications/notification-type.ts
export enum NotificationType {
  ADMIN_GROUP_CHANGED = 'ADMIN_GROUP_CHANGED',
  ADMIN_USER_CHARACTERISTICS_CHANGED = 'ADMIN_USER_CHARACTERISTICS_CHANGED',
  ADMIN_USER_CREATED = 'ADMIN_USER_CREATED',
  ADMIN_USER_GROUP_CHANGED = 'ADMIN_USER_GROUP_CHANGED',
  ADMIN_USER_ROLE_CHANGED = 'ADMIN_USER_ROLE_CHANGED',
  ADMIN_USER_STATUS_CHANGED = 'ADMIN_USER_STATUS_CHANGED',
  TICKET_CREATED = 'TICKET_CREATED',
  TICKET_ASSIGNED = 'TICKET_ASSIGNED',
  TICKET_COMMENTED = 'TICKET_COMMENTED',
  TICKET_SLA = 'TICKET_SLA',
  TICKET_STATUS_CHANGED = 'TICKET_STATUS_CHANGED',
}